+++ Sora/Signaling.swift
@@ -46,15 +46,19 @@ private func updateMetadata(signaling: Signaling, data: Data) -> Signaling {
         json = jsonObject as! [String: Any]
     } catch {
         // JSON のシリアライズが失敗した場合は、引数として渡された signaling をそのまま返し、処理を継続する
-        Logger.info(type: .signaling,
+        Logger.error(type: .signaling,
                   message: "updateMetadata failed. error: \(error.localizedDescription), data: \(data)")
         return signaling
     }
 
     switch signaling {
     case .offer(var message):
+        // TODO: if json.keys.contains("key") を if let に書き換えたい
         if json.keys.contains("metadata") {
             message.metadata = json["metadata"]
+        }
+        if let dataChannels = json["data_channels"] as? [[String: Any]] {
+            message.dataChannels = dataChannels
         }
         return .offer(message)
     case .push(var message):
@@ -120,10 +124,13 @@ public enum Signaling {
     case pong(SignalingPong)
     
     /// "disconnect" シグナリング
-    case disconnect
+    case disconnect(SignalingDisconnect)
     
     /// "pong" シグナリング
     case push(SignalingPush)
+    
+    /// "switch" シグナリング
+    case switched(SignalingSwitched)
     
     /// :nodoc:
     public static func decode(_ data: Data) -> Result<Signaling, Error> {
@@ -164,6 +171,8 @@ public enum Signaling {
             return "disconnect"
         case .push(_):
             return "push"
+        case .switched(_):
+            return "switched"
         }
     }
     
@@ -359,6 +368,9 @@ public struct SignalingConnect {
 
     /// :nodoc:
     public var environment: String?
+    
+    public var dataChannelSignaling: Bool?
+    public var ignoreDisconectWebSocket: Bool?
 
 }
 
@@ -440,6 +452,9 @@ public struct SignalingOffer {
 
     /// エンコーディング
     public let encodings: [Encoding]?
+    
+    /// データ・チャンネルの設定
+    public var dataChannels: [[String: Any]] = []
 
 }
 
@@ -500,6 +515,13 @@ public struct SignalingPush {
     /// プッシュ通知で送信される JSON データ
     public var data: Any? = {}
     
+}
+
+/**
+ "switched" シグナリングメッセージを表します。
+ */
+public struct SignalingSwitched {
+    public var ignoreDisconnectWebSocket: Bool?
 }
 
 /**
@@ -751,6 +773,10 @@ public struct SignalingPing {
  */
 public struct SignalingPong {}
 
+public struct SignalingDisconnect {
+    public var reason: String?
+}
+
 // MARK: -
 // MARK: Codable
 
@@ -795,6 +821,8 @@ extension Signaling: Codable {
             self = .push(try SignalingPush(from: decoder))
         case "re-offer":
             self = .reOffer(try SignalingReOffer(from: decoder))
+        case "switched":
+            self = .switched(try SignalingSwitched(from: decoder))
         default:
             throw SoraError.unknownSignalingMessageType(type: type)
         }
@@ -823,8 +851,9 @@ extension Signaling: Codable {
             try message.encode(to: encoder)
         case .pong:
             try container.encode(MessageType.pong.rawValue, forKey: .type)
-        case .disconnect:
+        case .disconnect(let message):
             try container.encode(MessageType.disconnect.rawValue, forKey: .type)
+            try message.encode(to: encoder)
         default:
             throw SoraError.invalidSignalingMessage
         }
@@ -916,6 +945,8 @@ extension SignalingConnect: Codable {
         case sora_client
         case libwebrtc
         case environment
+        case data_channel_signaling
+        case ignore_disconnect_websocket
     }
     
     enum VideoCodingKeys: String, CodingKey {
@@ -947,7 +978,9 @@ extension SignalingConnect: Codable {
         try container.encodeIfPresent(soraClient, forKey: .sora_client)
         try container.encodeIfPresent(webRTCVersion, forKey: .libwebrtc)
         try container.encodeIfPresent(environment, forKey: .environment)
-
+        try container.encodeIfPresent(dataChannelSignaling, forKey: .data_channel_signaling)
+        try container.encodeIfPresent(ignoreDisconectWebSocket, forKey: .ignore_disconnect_websocket)
+        
         if videoEnabled {
             if videoCodec != .default || videoBitRate != nil {
                 var videoContainer = container
@@ -1280,6 +1313,37 @@ extension SignalingPong: Codable {
     
     public func encode(to encoder: Encoder) throws {
         // エンコードするプロパティはない
+    }
+    
+}
+
+/// :nodoc:
+extension SignalingDisconnect: Codable {
+    
+    enum CodingKeys: String, CodingKey {
+        case reason
+    }
+    
+    public init(from decoder: Decoder) throws {
+        throw SoraError.invalidSignalingMessage
+    }
+    
+    public func encode(to encoder: Encoder) throws {
+        var container = encoder.container(keyedBy: CodingKeys.self)
+        try container.encode(reason, forKey: .reason)
+    }
+
+}
+
+extension SignalingSwitched: Decodable {
+    
+    enum CodingKeys: String, CodingKey {
+        case ignore_disconnect_websocket
+    }
+    
+    public init(from decoder: Decoder) throws {
+        let container = try decoder.container(keyedBy: CodingKeys.self)
+        ignoreDisconnectWebSocket = try container.decode(Bool.self, forKey: .ignore_disconnect_websocket)
     }
     
 }